load waterfall chart on its own URL
