+++ frontend/src/scenes/performance/WebPerformance.scss
@@ -4,7 +4,7 @@
     max-width: 300px;
 }
 
-.performance-waterfall {
+.web-performance {
     line-height: 32px;
     padding-right: 3em;
